requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
pydantic==2.11.7
python-dotenv==1.1.1
httpx==0.28.1
python-jose[cryptography]==3.5.0